seethos-api-server/requirements_utf8.txt
aiohttp==3.8.5
aiosignal==1.3.1
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
billiard==4.1.0
boto3==1.28.62
botocore==1.31.62
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.4
defusedxml==0.8.0rc2
dj-rest-auth==5.0.1
Django==4.2.5
django-allauth==0.57.0
django-ckeditor==6.5.1
django-cors-headers==3.3.0
django-extensions==4.1
django-filter==23.3
django-js-asset==2.1.0
django-rest-framework==0.1.0
django-sendgrid-v5==1.2.2
django-storages==1.14.2
django-templated-mail==1.1.1
django-tinymce==3.5.0
djangorestframework==3.14.0
djangorestframework-swagger==0.2.2
djoser==2.2.0
dnspython==2.4.2
drf-yasg==1.21.7
frozenlist==1.4.0
idna==3.4
inflection==0.5.1
jmespath==1.0.1
kombu==5.3.2
multidict==6.0.4
oauthlib==3.2.2
openai==1.7.0
packaging==23.2
prompt-toolkit==3.0.36
psycopg2-binary
pycparser==2.21
pymongo
python-dateutil==2.8.2
python-dotenv==1.0.0
python-http-client==3.3.7
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
s3transfer==0.7.0
sendgrid==6.10.0
sentry-sdk==1.31.0
six==1.16.0
social-auth-app-django==5.3.0
social-auth-core==4.4.2
sqlparse==0.4.4
starkbank-ecdsa==2.2.0
stripe==6.6.0
tqdm==4.66.1
typing_extensions==4.8.0
tzdata==2023.3
uritemplate==4.1.1
vine==5.0.0
wcwidth==0.2.8
yarl==1.9.2
PyPDF2
python_docx
openpyxl
pandas
celery
redis~=5.0.1
demjson3
sentry-sdk